Thomson module restructured
The thomson module was modified to fit the new structure of the package. No modifications regarding function.
