+++ __init__.py
@@ -1,4 +1,6 @@
 from .get_data_gpi import get_data_gpi, add_coordinate_gpi
+from .get_data_thomson import get_data_thomson, add_coordinate_thomson
+
 from .get_data import register
 
 from . import gpi
