+++ thomson/__init__.py
@@ -1,1 +1,6 @@
-from .calculate_gpi_thomson_elm_results import 
+from .fit_nstx_thomson_profiles import get_fit_nstx_thomson_profiles
+from .calculate_gpi_thomson_elm_results import get_all_thomson_data_for_elms, get_elms_with_thomson_profile, plot_elm_properties_vs_gradient
+from .calculate_gpi_thomson_elm_results import plot_elm_parameters_vs_ahmed_fitting, plot_elm_properties_vs_max_gradient
+from .calculate_gpi_thomson_elm_results import plot_elm_properties_vs_gradient_before_vs_after
+
+from .get_nstx_thomson_gradient import get_nstx_thomson_gradient
