+++ thomson/calculate_gpi_thomson_elm_results.py
@@ -17,9 +17,8 @@ import pickle
 import flap
 import flap_nstx
 
-from flap_nstx.analysis import calculate_nstx_gpi_avg_frame_velocity, calculate_nstx_gpi_smooth_velocity
-from flap_nstx.analysis import flap_nstx_thomson_data, get_nstx_thomson_gradient, get_fit_nstx_thomson_profiles
-from flap_nstx.analysis import read_ahmed_fit_parameters
+from flap_nstx.thomson import get_nstx_thomson_gradient, get_fit_nstx_thomson_profiles
+from flap_nstx.publications import read_ahmed_fit_parameters
 
 from matplotlib.backends.backend_pdf import PdfPages
 thisdir = os.path.dirname(os.path.realpath(__file__))
@@ -42,7 +41,7 @@ def get_all_thomson_data_for_elms():
         previous_shot=shot
         start_time=time.time()
         try:
-            a=flap_nstx_thomson_data(shot, force_mdsplus=True)
+            flap.get_data('NSTX_THOMSON', exp_id=shot, options={'force_mdsplus':True})
         except:
             failed_shot.append({'Shot':shot})
             n_fail_shots+=1.
